[Global Hist] Enabling Global Histogram Property for CRTC
This commit enables GLOBAL_HIST_EN DRM property for connected CRTC if,
Connected panel is of eDP type.
Currently we are enabling it on loading mutter, but going forward we
are working on to give an GUI option inside Settings-->Displays.
That work is in progress and will integrated soon once we are done.
Sign-Off: Mayani, Smit <smit.mayani@intel.com>