[DPST Algo] DPST Algo integration with Global Histogram in mutter
This commit is to integrate DPST algorithm along with Global Histogram
components both in  Kernel and Mutter.
So when we want to pass Histogram data along with Resolutino and Aggressiveness
to Algo API, this commit uses internal DPST Algorithm to get Global_IET and Backlight
calculated based on the input provided to the library.
Sign-Off: Mayani, Smit <smit.mayani@intel.com>